+++ CHANGES.md
@@ -1,6 +1,17 @@
 CHANGES
 =======
 
+- `pom.xml`
+    * Updated the version of authlete-java-common from 4.27 to 4.33.
+
+- `AuthleteApiCaller` class
+    * Added the `callCredentialNonce(CredentialNonceRequest, Options)` method.
+
+- New types
+    * `BaseCredentialNonceEndpoint` class
+    * `CredentialNonceRequestHandler` class
+
+
 2.89 (2025-11-04)
 -----------------